# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/search": {
            "get": {
                "description": "Get Companies by term",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get Companies by term",
                "parameters": [
                    {
                        "type": "string",
                        "description": "term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Enum: 0 -\u003e Account, 1 -\u003e Name, 2 -\u003e Address, 3 -\u003e ManagerName, 4 -\u003e MangerTel, 5 -\u003e BusinessLicenseNumber",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company Info",
                        "schema": {
                            "$ref": "#/definitions/response.GetCompaniesResp"
                        }
                    }
                }
            }
        },
...
        "/manage/company/info": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Allow Update For Companies",
                "parameters": [
                    {
                        "description": "allow companies update request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AllowCompaniesUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "allow update for companies response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/manage/company/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Allow Registration For Companies",
                "parameters": [
                    {
                        "description": "array of companies' accounts",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompaniesRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "allow companies response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/manage/company/reject-registration": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Reject Registration For Companies",
                "parameters": [
                    {
                        "description": "reject companies register request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompaniesRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "reject companies response"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "company_model.CompanyInfo": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "联创"
                },
                "address": {
                    "type": "string",
                    "example": "启明学院亮胜楼八楼"
                },
                "business_license_number": {
                    "type": "string",
                    "example": "114514"
                },
                "manager_name": {
                    "type": "string",
                    "example": "汉堡"
                },
                "manager_tel": {
                    "type": "string",
                    "example": "1919810"
                },
                "name": {
                    "type": "string",
                    "example": "联小创在线科技有限公司"
                }
            }
        },
...
        "request.AllowCompaniesUpdateReq": {
            "type": "object",
            "properties": {
                "company_accounts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "request.CompaniesRegisterReq": {
            "type": "object",
            "properties": {
                "company_accounts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.GetCompaniesResp": {
            "type": "object",
            "properties": {
                "company_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/company_model.CompanyInfo"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}